# FinalProject/FinalProject.web/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "FinalProject.web, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/MyDay": {
      "get": {
        "tags": [
          "MyDay"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MyTask"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MyTask"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MyTask"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "MyDay"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MyTask"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MyTask"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MyTask"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "MyDay"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MyTask"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MyTask"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MyTask"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/MyDay/{name}": {
      "get": {
        "tags": [
          "MyDay"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MyTask"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyTask"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyTask"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "MyDay"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/MyDay/Done": {
      "post": {
        "tags": [
          "MyDay"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MyTask"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MyTask"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MyTask"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/MyDay/Deletebystatus": {
      "delete": {
        "tags": [
          "MyDay"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MyTask"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MyTask"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MyTask"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "MyTask": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "boolean"
          },
          "dueDate": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      }
    }
  }
}